Remove `schedulerStrategy: pack | spread`
This was a half measure solution in order to pack user pods on nodes for 
better cluster autoscaling. The to be implemented custom scheduler will 
allow actual packing of user pods while this option that enables a 
preferred pod affinty made pods schedule on a nodes one or more user pod 
on it already.

For example, consider having three nodes: one empty node and two nodes 
with one user pod on them. A custom user scheduler could schedule five 
additional user pods to the same node by looking how much resoureces 
were requested, but simply adding a user pod affinity would only ensure 
that new pods are scheduled on one of the two nodes with one or more 
user pod on it. This option is overpromising things, let us not mislead 
the users and direct them towards a custom scheduler instead.