Image-puller's made robost (Helm 2.9+)
The hook based puller ensures user images are pulled to the nodes before 
an upgrade of the hub takes place. This is helpful as if this wouldn't 
be done, users would potentially be spawned on nodes without the user 
image available forcing them to wait up to minutes. The users would 
probably think something was wrong and shut down ahead of time.

To complement this pre-upgrade / pre-install hook based image puller, 
the continuous puller will pull images to all nodes right when they are 
created, for example by a cluster autoscaler. The continuous image 
puller is now enabled by default as it is very useful for the cluster 
autoscaling used in conjunction with user placehodlers that is creating 
headroom.

Due to a bug in Helm related to the annotation 
`helm.sh/hook-delete-policy: hook-failed`, multiple puller instances 
tended to accumulate. In this commit the deletion policy 
`before-hook-creation` enters our code base (Helm 2.9+) and allows for a 
pre-existing k8s resource with the same name to be deleted before helm 
tries to create it. Without this Helm would raise an "... already exist" 
error. This also ensures that max one pre-install/upgrade hook puller 
would remain on install/upgrade failures. It also ensures that this 
single resource as well is cleaned up on the next successfull helm 
installation or upgrade.